fix(liveslots): cache.delete() does not return a useful value
Liveslots uses an internal `Cache` utility to hold data about virtual
objects, backed by the vatstore. PR #8752 included a change to make
its `delete()` method return a "did exist" boolean, just like a
JavaScript `Map`. Unfortunately the cache does not know whether a key
is present/absent in the backing store, so `delete()` will return an
erroneous value. It would cost an extra DB query to make this behave
as expected, and liveslots does not have a need for it.

So this commit reverts that change, and makes `delete()` return void
as before.