refactor(kmarshal): simplify makeStandinPromise (#9475)
closes: #9434 
refs: #XXXX

## Description

This removes a kludge needed to accommodate an ancient version of endo. The kludge should no longer be needed, and the remaining code can be simpler.

However, we cannot actually merge this PR to master until we know we do not need to worry about this ancient version of endo.

- [ ] identify what version of endo fixed the problem that the kludge addressed
- [ ] figure out whether that should still inhibit merging this simplification.

Altogether, the kludge is small and cheap, so there's nothing urgent about merging this PR. We should wait until we're confident about the compat risk.

### Security Considerations

none
### Scaling Considerations

none
### Documentation Considerations

none
### Testing Considerations

none
### Upgrade Considerations

none