fix(liveslots): move passStyleOf from vatGlobals to inescapableGlobalProperties
Previously, liveslots added the special `PassStyleOfEndowmentSymbol`
on `vatGlobals`. This enabled the first-level vat bundle (the one that
defines buildRootObject, e.g. ZCF) to use the same passStyleOf as
liveslots is using, with a real WeakMap for its memoization cache.

However, child Compartments of that vat, such as the ones contract
bundles are loaded into, do not automatically get the same global, and
ZCF does not pass any special options to `importBundle()`, which would
enable that.

This commit attaches the symbol to `inescapableGlobalProperties`,
rather than `vatGlobals`. Those properties are automatically (indeed
unavoidably) copied onto the `globalThis` of each child Compartment,
recursively. This is stronger than we need: we're ok with a
Compartment choosing to omit or change this property.

This allows contract code which does `import { passStyleOf } from
'@endo/pass-style'` to get the correct/fast/cheap version too.

Tests were added to liveslots to assert that it gives the expected
`inescapableGlobalProperties` option when asking the supervisor to
build the vat's namespace, and to SwingSet to assert that both child
and grandchild Compartments get the expected version of
passStyleOf (which was not the case for mere `vatGlobals`).

fixes #9981