ci: drop xs:unit job
There are few to no errors we've detected in unit tests running in XS that we didn't in Node. We now have many other tests that run XS and don't need to unit test these. It generally doesn't hurt to test more, but ava-xs is so different from XS that the cost of reconciliation is too great.