chore(swingset/misc-tools): add #9039 scanner to count cleanup work
Give this tool a swingstore DB, and it will identify all the vat/kpid
pairs that need cleanup to remediate the #9039 bug. This also
estimates the amount of work the kernel-side remediation tool will
need to do the cleanup automatically.

The tool needed 11,882 DB queries to find the 3166 KPIDs needing
cleanup in run-53, and ran in 0.37s on my laptop