chore(swingset): add a proper PromiseRecord type
Improve the types we use to represent records in the kernel promise
table, and the values we pass around when working with that table.

Update the kernel code (mainly kernelKeeper) to satisfy the new
variants, by constructing the records with more care. Several
assert/Fail lines were changed to satisfy the typechecker.

As a side effect, this fixes #9889, because the casual/buggy approach
we took before was correctly identified by TSC as not working for all
promise states.

fixes #9889