build proposals for PriceFeed coreEvals on multiple platforms. (#10146)
closes: #10076
refs: #10074

## Description

Build platform-dependent proposals for the PriceFeed coreEval. 

For each platform (not counting A3P) defined in `updatePriceFeeds.js`, a command is added, which invokes `scripts/build-submission.sh` three times to build each of the required proposals (updatePriceFeeds.js, add-auction.js, and upgradeVaults.js). The invocation for updatePriceFeeds takes a parameter specifying the platform.

### Security Considerations

N/A

### Scaling Considerations

N/A

### Documentation Considerations

Not user-visible

### Testing Considerations

test on DevNet and MainFork

### Upgrade Considerations

It's all about upgrade. This coreEval has code that varies depending on the collaterals and oracles present, so we have to build separate proposals for each platform.