fix(swingset): add remediation of 9039
We use the `upgradeSwingset()` mechanism (when switching to version 3)
to identify and fix all the leftover kpids caused by #9039. This
performs automatic remediation of the damage, exactly once per
database.

The v2-to-v3 upgrader will look for all c-list promise entries, then
check to see if the kpid is settled. It ignores the ones that have
`dispatch.notify` events in either the runQueue or the
acceptanceQueue. It then enqueues notify events to the vats, adjusts
the refcounts and kernel stats accordinging, and finally and bumps the
version to 3.

These kpids will be for promises which were decided by an upgraded
vat, but not resolved by upgrade time. The kernel "disconnects" (rejects)
these promises, but #9039 failed to remove them from the upgraded
vat's c-list.

On the first run after this remediation, the kernel will deliver those
notifies, which will delete the c-list entries and decrement their
refcounts, which may trigger the usual deletions and further
decrefs. The notifies will be delivered to the vat's new incarnation,
which will harmlessly ignore them (as unrecognized vpids). This turned
out to be the easiest way to get all the possible cleanups to run.

New tests were added to exercise the remediation code.

As this implements the last part of the fix, it:

fixes #9039